Workflow.OSX.DSL

Description

some higher-level workflows you can derive from the primitives in Workflow. (also see the source)

Documentation

delay :: forall m. MonadFree WorkflowF m => Time -> m ()

openURL :: forall m. MonadFree WorkflowF m => URL -> m ()

openApplication :: forall m. MonadFree WorkflowF m => Application -> m ()

currentApplication :: forall m. MonadFree WorkflowF m => m Application

setClipboard :: forall m. MonadFree WorkflowF m => ClipboardText -> m ()

getClipboard :: forall m. MonadFree WorkflowF m => m ClipboardText

sendText :: forall m. MonadFree WorkflowF m => String -> m ()

sendKeyChord :: forall m. MonadFree WorkflowF m => [Modifier] -> Key -> m ()

insert :: MonadWorkflow m => String -> m ()

copy :: MonadWorkflow m => m String

access the currently selected region from Haskell, via the clipboard

paste :: MonadWorkflow m => m ()

google :: MonadWorkflow m => String -> m ()

google a query. properly encodes the url.