Workflow.OSX.Extra

Description

assorted functionality, imported by most modules in this package.

Documentation

failed :: MonadThrow m => String -> m a

module Data.Data

module GHC.Generics

(<>) :: Monoid m => m -> m -> m infixr 6

An infix synonym for mappend.

Since: 4.5.0.0

traverse_ :: (Foldable t, Applicative f) => (a -> f b) -> t a -> f ()

Map each element of a structure to an action, evaluate these actions from left to right, and ignore the results. For a version that doesn't ignore the results see traverse.